🗃 Added fingerprint model

+++ schedulesy/apps/ade_api/refresh.py
@@ -1,6 +1,7 @@
 import json
+import time
 
-from schedulesy.models import Resource
+from schedulesy.models import Resource, Fingerprint
 from .ade import ADEWebAPI, Config
 
 class Flatten:
@@ -44,16 +45,36 @@ class Refresh:
         myade = ADEWebAPI(**config)
         myade.connect()
         myade.setProject(5)
+        method = 'getResources'
+        self.data = {}
         for r_type in ['classroom', 'instructor', 'trainee', 'category5']:
-            Resource.objects.all().filter(fields__category=r_type).delete()
             tree = myade.getResources(category=r_type, detail=3, tree=True, hash=True)
-            open("{}.json".format(r_type), 'w').write(json.dumps(tree))
-            print(tree['hash'])
-            test = Flatten(tree['data']).f_data
+            n_fp = tree['hash']
 
-            for key, value in test.items():
-                resource = Resource(ext_id=key, fields=value)
-                resource.save()
+            try:
+                o_fp = Fingerprint.objects.all().get(ext_id=r_type, method=method)
+            except:
+                o_fp = None
 
-        self.data = test
-        print("done")
+            key = "{}-{}".format(method, r_type)
+            self.data[key] = {'status': 'unchanged', 'fingerprint': n_fp}
+
+            if not o_fp or o_fp.fingerprint != n_fp:
+                start = time.clock()
+                Resource.objects.all().filter(fields__category=r_type).delete()
+                test = Flatten(tree['data']).f_data
+
+                count = 0
+                for k, v in test.items():
+                    resource = Resource(ext_id=k, fields=v)
+                    resource.save()
+                    count += 1
+                if o_fp:
+                    o_fp.fingerprint = n_fp
+                else:
+                    o_fp=Fingerprint(ext_id=r_type, method=method, fingerprint=n_fp)
+                o_fp.save()
+                elapsed = time.clock() - start
+                self.data[key]['status'] = 'updated'
+                self.data[key]['count'] = count
+                self.data[key]['elapsed'] = elapsed
